requirements.txt
tensorflow>=2.10.0
opencv-python>=4.6.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
seaborn>=0.11.0
Pillow>=9.0.0 